backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
Pillow>=10.3.0
opencv-python-headless>=4.9.0
anthropic>=0.40.0
numpy>=1.26.0
cryptography>=42.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
pydantic>=2.0.0
